Entity 필드 자료형 변경

+++ Dashboard/src/main/java/runday/domain/LogCreated.java
@@ -11,6 +11,6 @@ public class LogCreated extends AbstractEvent {
     private Long id;
     private String userId;
     private Date startTime;
-    private Long startLatitude;
-    private Long startLongitude;
+    private Float startLatitude;
+    private Float startLongitude;
 }
